@@ -7,29 +7,33 @@
   "main": "main.js",
   "dependencies": {
     "big.js": "^5.0.3",
-    "expo": "16.0.0",
-    "react": "16.0.0-alpha.6",
-    "react-moment": "0.2.4",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-17.0.0.tar.gz",
-    "react-native-country-picker-modal": "^0.3.0-rc1",
-    "react-native-device-info": "^0.10.2",
-    "react-native-elements": "0.13.0",
+    "expo": "24.0.0",
+    "fbjs": "0.8.16",
+    "moment-timezone": "^0.5.11",
+    "native-base": "^2.3.5",
+    "prop-types": "^15.6.0",
+    "react": "16.0.0-beta.5",
+    "react-moment": "0.6.8",
+    "react-native": "github:expo/react-native#sdk-24.0.0",
+    "react-native-country-picker-modal": "^0.5.1",
+    "react-native-device-info": "^0.13.0",
+    "react-native-elements": "0.18.5",
     "react-native-infinite-scroll-view": "^0.4.2",
     "react-native-loading-spinner-overlay": "^0.5.0",
-    "react-native-modal": "^2.4.0",
+    "react-native-modal": "^4.1.1",
     "react-native-picker-dropdown": "0.0.1",
-    "react-native-popup-dialog": "^0.9.32",
-    "react-native-vector-icons": "^4.1.1",
-    "react-navigation": "git+https://github.com/react-community/react-navigation.git"
+    "react-native-popup-dialog": "^0.10.44",
+    "react-native-vector-icons": "^4.3.0",
+    "react-navigation": "^1.0.0-beta.21"
   },
   "devDependencies": {
-    "babel-eslint": "^7.2.3",
-    "eslint": "^3.19.0",
+    "babel-eslint": "^8.0.3",
+    "eslint": "^4.13.1",
     "eslint-config-exponent": "^5.1.3",
     "eslint-plugin-babel": "^4.1.1",
     "eslint-plugin-flowtype": "^2.32.1",
     "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-react": "^6.10.3"
+    "eslint-plugin-react": "^7.5.1"
   },
   "eslintConfig": {
     "extends": "exponent/react",
