@@ -1,6 +1,7 @@
 import React from 'react'
 import { TouchableOpacity } from 'react-native'
 import Icon from 'react-native-vector-icons/Ionicons'
+import PropTypes from 'prop-types';
 
 const DrawerButton = ({ navigation }) => (
   <TouchableOpacity style={{padding: 10}} underlayColor={'white'}>
@@ -14,7 +15,7 @@ const DrawerButton = ({ navigation }) => (
 )
 
 DrawerButton.propTypes = {
-  navigation: React.PropTypes.object.isRequired,
+  navigation: PropTypes.object.isRequired,
 };
 
 export default DrawerButton
